requirements.txt
fastapi
uvicorn[standard]
alembic
structlog
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
httpx
google-auth
requests
